backend/connectors/requirements.txt
aiogram==3.3.0
httpx==0.25.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
